small fix in test

+++ test/acceptance/schema/default_compression_test.go
@@ -118,7 +118,7 @@ func TestDefaultCompressionRQ8(t *testing.T) {
 	jsonRescoreLimit := rq.(map[string]interface{})["rescoreLimit"].(json.Number)
 	rescoreLimit, err := jsonRescoreLimit.Int64()
 	require.Nil(t, err)
-	require.Equal(t, hnsw.DefaultRQRescoreLimit, rescoreLimit)
+	require.Equal(t, int64(hnsw.DefaultRQRescoreLimit), rescoreLimit)
 	skipDefaultQuantization := viconfig.(map[string]interface{})["skipDefaultQuantization"].(bool)
 	require.Equal(t, false, skipDefaultQuantization)
 	trackDefaultQuantization := viconfig.(map[string]interface{})["trackDefaultQuantization"].(bool)
